Moving another API into the /v1/accounts namespace.
